backend/requirements.txt
# Core AI dependencies - Agnóstico para qualquer provedor
google-generativeai>=0.3.0  # Para Google Gemini (pode ser substituído)
python-dotenv>=0.19.0

# Logging e utilities
loguru>=0.6.0

# Protocolos A2A e MCP - Dependências opcionais
pydantic>=1.8.0  # Para validação de dados dos protocolos
requests>=2.25.0  # Para comunicação HTTP entre agentes (opcional)
websockets>=10.0  # Para comunicação WebSocket em tempo real (opcional)

# Nota: uuid, datetime, enum e typing são built-in no Python 3.6+
# Nota: sqlite3 é built-in no Python padrão